refactor: extract PredictionMarketsPage components + formatters (F5)
Extract the four presentational components (ProbabilityBar,
AnalysisLoadingPanel, AnalysisPanel, MarketCard) plus their component-only
config (STANCE_META, ANALYSIS_LOADING_STEPS, DEFAULT_ANALYSIS_STATE,
getAnalysisState) into components/prediction-markets/components.js, and
the formatters into components/prediction-markets/format.js. The page
keeps its market list state/fetch logic and renders MarketCard, dropping
from 799 to 376 lines. Completes the F5 breakup of the seven >1000-line
pages. Pure code movement, guarded by the existing 5-test suite; lint +
build clean.
